[system test] Move TestUtils methods used only in STs to systemtest module
